added a slide show container
the slide show container uses java script ,css and html.
currently there is a 1/3 caption on top left side of the
slide show container, which need to be removed.
captions at the bottome middle needs to be removed
can move slide show with arrows on the sides or dots on the bottom