added product page and implemented some of the backend
created a directory called backend that will store the nodejs/express.js server
that will be made for the data base. created a whole schema that will be used with
the main site.tested database with postman to make sure things lined up fine.
will need some more fine tuning in the furture. product page is a simple
page with that shows off the product this will need some work as well