added to backend and frontend
added cors to the backend so the website can communicate with the nodejs server without error.
Added front page products as a table to the back end then made post and get request for the api to handle this.
I then overhauled the products portion of the front page to list off all products in the front page products table.
Added 7 new files to the app 4 being services the app will use or will need.
3 being models for the product, orders and user portion of the app.
This allows for encapsulation between the backend and the front end.