added products to product page
added products to product page tried to add pagination
to the point where only 12 products are displayed on the
screen at a time. currently it displays all. need to
work out kinks to get pagination working properly
added functionality to the backend such as get total
number of product and grab a range of products in prep
for pagination.